refactor: replace video feed components with GenericVideoFeed for improved code reuse
